+++ src/Autocomplete.tsx
@@ -1,0 +1,95 @@
+import React, { useCallback, useMemo, useState } from 'react';
+import { Person } from './types/Person';
+import debounce from 'lodash.debounce';
+
+type Props = {
+  people: Person[];
+  delay?: number;
+  onSelected: (person: Person | null) => void;
+};
+
+export const Autocomplete: React.FC<Props> = ({
+  people,
+  delay = 300,
+  onSelected,
+}) => {
+  const [query, setQuery] = useState('');
+  const [isActive, setIsActive] = useState(false);
+  const [appliedQuery, setAppliedQuery] = useState('');
+  const [selectedPerson, setSelectedPerson] = useState<Person | null>(null);
+
+  const applyQuery = useCallback(debounce(setAppliedQuery, delay), []);
+
+  const handleQueryChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setQuery(event.target.value);
+    applyQuery(event.target.value);
+    setIsActive(true);
+
+    if (selectedPerson && event.target.value !== selectedPerson.name) {
+      setSelectedPerson(null);
+      onSelected(null);
+    }
+  };
+
+  const filteredPeople = useMemo(() => {
+    return people.filter(person =>
+      person.name.toLowerCase().includes(appliedQuery.toLowerCase()),
+    );
+  }, [appliedQuery, query]);
+
+  return (
+    <div className="dropdown is-active">
+      <div className="dropdown-trigger">
+        <input
+          type="text"
+          placeholder="Enter a part of the name"
+          className="input"
+          data-cy="search-input"
+          value={query}
+          onChange={handleQueryChange}
+          onFocus={() => setIsActive(true)}
+          onBlur={() => setIsActive(false)}
+        />
+      </div>
+      {isActive && (
+        <div className="dropdown-menu" role="menu" data-cy="suggestions-list">
+          <div className="dropdown-content">
+            {filteredPeople.map(person => {
+              return (
+                <div
+                  className="dropdown-item"
+                  style={{ cursor: 'pointer' }}
+                  data-cy="suggestion-item"
+                  key={person.slug}
+                  onMouseDown={() => {
+                    setQuery(person.name);
+                    setIsActive(false);
+                    onSelected(person);
+                    setSelectedPerson(person);
+                  }}
+                >
+                  <p className="has-text-link">{person.name}</p>
+                </div>
+              );
+            })}
+          </div>
+        </div>
+      )}
+      {filteredPeople.length === 0 && (
+        <div
+          className="
+                notification
+                is-danger
+                is-light
+                mt-3
+                is-align-self-flex-start
+              "
+          role="alert"
+          data-cy="no-suggestions-message"
+        >
+          <p className="has-text-danger">No matching suggestions</p>
+        </div>
+      )}
+    </div>
+  );
+};
